# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/poke-lounge/rom-data": {
      "get": {
        "operationId": "PokeLoungeController_getRomData",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "data"],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [true]
                    },
                    "data": {
                      "$ref": "#/components/schemas/PokeLoungeRomDataResponseDto"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "필수 ROM 문서가 없거나 무결하지 않음"
          }
        },
        "summary": "ROM에서 추출한 Poke Lounge 게임 데이터 조회",
        "tags": ["poke-lounge"]
      }
    },
    "/poke-lounge/shops/basic/items": {
      "get": {
        "operationId": "PokeLoungeController_getBasicShopItemIds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "data"],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [true]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "minimum": 1
                      },
                      "minItems": 1,
                      "uniqueItems": true
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "일반 상점 판매 목록이 없거나 무결하지 않음"
          }
        },
        "summary": "Poke Lounge 일반 상점 판매 목록 조회",
        "tags": ["poke-lounge"]
      }
    },
    "/poke-lounge/shops/premium/items": {
      "get": {
        "operationId": "PokeLoungeController_getPremiumShopItemIds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "data"],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [true]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "integer",
                        "minimum": 1
                      },
                      "minItems": 1,
                      "uniqueItems": true
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "희귀 상점 판매 목록이 없거나 무결하지 않음"
          }
        },
        "summary": "Poke Lounge 희귀 상점 판매 목록 조회",
        "tags": ["poke-lounge"]
      }
    },
    "/poke-lounge/rooms": {
      "post": {
        "operationId": "PokeLoungeController_createRoom",
        "parameters": [
          {
            "name": "If-Match-Revision",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "example": "0"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePokeLoungeRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "data"],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [true]
                    },
                    "data": {
                      "$ref": "#/components/schemas/PokeLoungeRoomResponseDto"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/PokeLoungeRoomConflictResponseDto"
                    },
                    {
                      "$ref": "#/components/schemas/PokeLoungeRoomFullResponseDto"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": ["poke-lounge"]
      }
    },
    "/poke-lounge/rooms/{roomCode}": {
      "get": {
        "operationId": "PokeLoungeController_getRoom",
        "parameters": [
          {
            "name": "roomCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "afterRevision",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "data"],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [true]
                    },
                    "data": {
                      "$ref": "#/components/schemas/PokeLoungeRoomResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["poke-lounge"]
      }
    },
    "/poke-lounge/rooms/{roomCode}/competitive-seat": {
      "post": {
        "operationId": "PokeLoungeController_bindCompetitiveSeat",
        "parameters": [
          {
            "name": "roomCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindCompetitiveSeatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "data"],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [true]
                    },
                    "data": {
                      "$ref": "#/components/schemas/CompetitiveAssignmentResponseDto"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Account authentication is disabled; anonymous play remains available."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["poke-lounge"]
      }
    },
    "/poke-lounge/rooms/{roomCode}/matches/{matchId}/actions": {
      "post": {
        "operationId": "PokeLoungeController_submitCompetitiveAction",
        "parameters": [
          {
            "name": "roomCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCompetitiveActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "data"],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [true]
                    },
                    "data": {
                      "$ref": "#/components/schemas/CompetitiveActionResponseDto"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Account authentication is disabled; anonymous play remains available."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["poke-lounge"]
      }
    },
    "/poke-lounge/rooms/{roomCode}/matches/{matchId}/session-actions": {
      "post": {
        "operationId": "PokeLoungeController_submitSessionCompetitiveAction",
        "parameters": [
          {
            "name": "roomCode",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "matchId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitSessionCompetitiveActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "data"],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [true]
                    },
                    "data": {
                      "$ref": "#/components/schemas/CompetitiveActionResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": ["poke-lounge"]
      }
    },
    "/poke-lounge/rooms/quick-play": {
      "post": {
        "operationId": "PokeLoungeController_quickPlay",
        "parameters": [
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinPokeLoungeRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["success", "data"],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [true]
                    },
                    "data": {
                      "$ref": "#/components/schemas/PokeLoungeRoomResponseDto"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PokeLoungeRoomFullResponseDto"
                }
              }
            }
          }
        },
        "tags": ["poke-lounge"]
      }
    },
...
  },
  "info": {
    "title": "Poke Lounge API",
    "description": "Poke Lounge API 문서입니다.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PokeLoungeRoomConflictResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "code": {
            "type": "string",
            "enum": ["POKE_LOUNGE_REVISION_CONFLICT", "POKE_LOUNGE_IDEMPOTENCY_CONFLICT"]
          },
          "message": {
            "type": "string",
            "example": "Poke Lounge room revision conflict"
          },
          "snapshot": {
            "$ref": "#/components/schemas/PokeLoungeRoomResponseDto"
          }
        },
        "required": ["statusCode", "code", "message", "snapshot"]
      },
      "PokeLoungeRoomFullResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "code": {
            "type": "string",
            "enum": ["POKE_LOUNGE_ROOM_FULL"]
          },
          "message": {
            "type": "string",
            "example": "Poke Lounge room is full"
          }
        },
        "required": ["statusCode", "code", "message"]
      },
...
      "PokeLoungeRomDataResponseDto": {
        "type": "object",
        "properties": {
          "documents": {
            "minItems": 4,
            "maxItems": 4,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PokeLoungeRomDocumentDto"
            }
          }
        },
        "required": ["documents"]
      },
      "CreatePokeLoungeRoomDto": {
        "type": "object",
        "properties": {
          "roomCode": {
            "type": "string",
            "example": "ABC234"
          },
          "playerId": {
            "type": "string",
            "example": "player-a"
          },
          "sessionId": {
            "type": "string",
            "example": "session-a"
          },
          "userId": {
            "type": "string",
            "example": "user-123"
          },
          "displayName": {
            "type": "string",
            "example": "Player A"
          },
          "roundDurationMs": {
            "type": "number",
            "example": 180000,
            "description": "Test-only override; production preparation is fixed at 180000ms"
          }
        },
        "required": ["sessionId"]
      },
      "BindCompetitiveSeatDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "example": "session-a",
            "maxLength": 256
          }
        },
        "required": ["sessionId"]
      },
...
      "SubmitCompetitiveActionDto": {
        "type": "object",
        "properties": {
          "assignmentRevision": {
            "type": "number",
            "minimum": 0
          },
          "turn": {
            "type": "number",
            "minimum": 0
          },
          "clientCommandId": {
            "type": "string",
            "format": "uuid"
          },
          "action": {
            "$ref": "#/components/schemas/CompetitiveActionDto"
          }
        },
        "required": ["assignmentRevision", "turn", "clientCommandId", "action"]
      },
...
    }
  }
}